+++ src/index.js
@@ -25,6 +25,9 @@ export { default as parseToAST  } from "./parser/parseToAST.js" ;
 export { default as stringify    } from "./serializer/stringify.js" ;
 export { default as stringifyAST } from "./serializer/stringifyAST.js" ;
 
+export { default as fromJSON     } from "./convert/fromJSON.js" ;
+export { default as toJSON       } from "./convert/toJSON.js" ;
+
 /**
  * Library version, kept in sync with `package.json`.
  * @type {string}